requirements.txt
absl-py==2.1.0
accelerate==0.15.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
beartype==0.14.1
bypy==1.8.5
cachetools==5.5.0
charset-normalizer==3.4.0
clean-fid==0.1.35
colour-demosaicing==0.1.6
colour-science==0.3.16
cycler==0.11.0
denoising-diffusion-pytorch==2.0.5
diffusers==0.15.1
dill==0.3.7
einops==0.6.0
ema-pytorch==0.4.8
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.2.2
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==2.10.0
huggingface-hub==0.16.4
idna==3.10
imageio==2.25.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==6.7.0
iopath==0.1.10
joblib==1.3.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kmeans-pytorch==0.3
kornia==0.6.12
lmdb==1.5.1
lpips==0.1.4
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mmengine==0.8.2
multidict==6.0.5
multiprocess==0.70.15
networkx==2.6.3
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.5.3.56
opencv-python-headless==4.1.2.30
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
platformdirs==4.0.0
portalocker==2.7.0
protobuf==3.20.1
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyDeprecate==0.3.2
Pygments==2.17.2
pyiqa==0.1.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-git==2018.2.1
pytorch-fid==0.3.0
pytorch-lightning==0.7.6
pytorch-msssim==1.0.0
pytz==2024.2
PyWavelets==1.3.0
PyYAML==5.4
rawpy==0.18.1
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.19.3
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.8.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
tabulate==0.9.0
tb-nightly==2.12.0a20230113
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.5
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2021.11.2
timm==0.9.7
tokenizers==0.13.3
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchgeometry==0.1.2
torchist==0.2.3
torchmetrics==0.11.4
torchstat==0.0.7
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.15.0
